Use unified coordinate utils
